chat/requirements.txt
aiohttp==3.7.4.post0
multidict==5.1.0
python-engineio==3.9.0
python-socketio==4.5.0
asgiref==3.3.1
Django==2.2
django-cors-headers==3.7.0
django-webpack-loader==0.7.0
djangorestframework==3.12.2
gunicorn==20.0.4
mysqlclient==2.0.3
psycopg2-binary==2.8.6
pytz==2020.5
sqlparse==0.4.1
requests==2.25.1
enum-compat==0.0.3
eventlet==0.30.0
greenlet
pytz==2020.5
six==1.10.0